requirements.txt
# 4コマ漫画自動生成システム 依存パッケージ

# Gemini API
google-generativeai>=0.3.0

# 画像処理
Pillow>=10.0.0

# YAML処理
PyYAML>=6.0

# その他
requests>=2.31.0
python-dateutil>=2.8.2
